+++ ui/src/com/kotcrab/vis/ui/widget/file/FileChooserStyle.java
@@ -40,6 +40,8 @@ public class FileChooserStyle {
 	public Drawable iconFilePdf;
 	public Drawable iconFileAudio;
 
+	public Drawable expandDropdown;
+
 	public FileChooserStyle () {
 	}
 }