+++ ui/src/com/kotcrab/vis/ui/widget/file/internal/FileSuggestionPopup.java
@@ -17,59 +17,41 @@
 package com.kotcrab.vis.ui.widget.file.internal;
 
 import com.badlogic.gdx.files.FileHandle;
-import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.ObjectMap;
 import com.kotcrab.vis.ui.widget.MenuItem;
-import com.kotcrab.vis.ui.widget.PopupMenu;
 import com.kotcrab.vis.ui.widget.VisTextField;
 import com.kotcrab.vis.ui.widget.file.FileChooser;
 
 /** @author Kotcrab */
-public class SuggestionPopupMenu extends PopupMenu {
-	private static final Vector2 tmpVector = new Vector2();
-	private static final int MAX_SUGGESTIONS = 10;
-
-	private final FileChooser chooser;
-
-	public SuggestionPopupMenu (FileChooser chooser) {
-		super(chooser.getChooserStyle().popupMenuStyleName);
-		this.chooser = chooser;
+public class FileSuggestionPopup extends AbstractSuggestionPopup {
+	public FileSuggestionPopup (FileChooser chooser) {
+		super(chooser);
 	}
 
-	public void fileNameKeyTyped (Stage stage, ObjectMap.Keys<FileHandle> files, VisTextField fileNameField) {
-		if (fileNameField.getText().length() == 0) {
+	public void pathFieldKeyTyped (Stage stage, ObjectMap.Keys<FileHandle> files, VisTextField pathField) {
+		if (pathField.getText().length() == 0) {
 			remove();
 			return;
 		}
 
-		int suggestions = createSuggestions(files, fileNameField);
+		int suggestions = createSuggestions(files, pathField);
 		if (suggestions == 0) {
 			remove();
 			return;
 		}
 
-		Vector2 pos = fileNameField.localToStageCoordinates(tmpVector.setZero());
-		float menuY;
-		if (pos.y - getHeight() <= 0) {
-			menuY = pos.y + fileNameField.getHeight() + getHeight() - 1;
-		} else {
-			menuY = pos.y + 1;
-		}
-		showMenu(stage, pos.x, menuY);
+		showMenu(stage, pathField);
 	}
 
 	private int createSuggestions (ObjectMap.Keys<FileHandle> files, VisTextField fileNameField) {
 		clearChildren();
 		int suggestions = 0;
 		for (final FileHandle file : files) {
-			if (file.name().startsWith(fileNameField.getText())) {
-				MenuItem item = new MenuItem(getCroppedName(file.name()));
-				item.getImageCell().size(0);
-				item.getShortcutCell().space(0).pad(0);
-				item.getSubMenuIconCell().size(0).space(0).pad(0);
+			if (file.name().startsWith(fileNameField.getText()) && file.name().equals(fileNameField.getText()) == false) {
+				MenuItem item = createMenuItem((getTrimmedName(file.name())));
 				item.addListener(new ChangeListener() {
 					@Override
 					public void changed (ChangeEvent event, Actor actor) {
@@ -88,7 +70,7 @@ public class SuggestionPopupMenu extends PopupMenu {
 		return suggestions;
 	}
 
-	private String getCroppedName (String name) {
+	private String getTrimmedName (String name) {
 		if (name.length() > 40) {
 			return name.substring(0, 40) + "...";
 		} else {